% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{append.info}
\alias{append.info}
\title{Merge selection coefficients, spatial concentration, conservation scores and gene expression data}
\usage{
append.info(output.folder, output.prefix)
}
\arguments{
\item{output.folder:}{A folder containing a "genes.opt.txt" file and a 'spatial.txt' file, a 'fitch.txt' file.}

\item{output.prefix:}{A prefix string used label the various files. Combined data frame is written to a file "prefix.combined.txt".}
}
\description{
Merge selection coefficients, spatial concentration, conservation scores and gene expression data
}
\examples{
: append.info(output.folder='./examples/', output.prefix='TCGA.ACC');
}
