% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{check.gene.consv.all}
\alias{check.gene.consv.all}
\title{Retreve fitch rate for each position of a protein.}
\usage{
check.gene.consv.all(id, type = "symbol")
}
\arguments{
\item{id:}{An ID indicating the gene.}

\item{type:}{A flag indicating what type the "id" is. Possible values are "symbol" (gene symbol), "gene.id" (Entrez gene ID), "ensembl.cds" (Ensemble transcript ID), "ensembl.protein" (Ensemble protein ID), and "ensembl.gene" (Ensembl gene ID).}
}
\value{
A vector containing fitch rate using 100 vertebrate alignments.
}
\description{
Retreve fitch rate for each position of a protein.
}
