% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{compute.fitch.stats}
\alias{compute.fitch.stats}
\title{Compute statistics related to evolutionary conservation}
\usage{
compute.fitch.stats(output.folder, output.prefix)
}
\arguments{
\item{output.folder:}{A folder containing a "mut.summary.txt" file and a 'spatial.txt' file.}

\item{output.prefix:}{A prefix string used label the various files. Conservation statistics is written to a file "prefix.fitch.txt".}
}
\description{
Compute statistics related to evolutionary conservation
}
\examples{
: compute.fitch.stats(output.folder='./examples/', output.prefix='TCGA.ACC');
}
