% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{compute.selection.stats}
\alias{compute.selection.stats}
\title{Compute statistics related to somatic selection}
\usage{
compute.selection.stats(output.folder = "", output.prefix = "")
}
\arguments{
\item{output.folder:}{A folder containing aggregated mutations produced by the parse.aggregated.mut() function.}

\item{output.prefix:}{Prefix used in the parse.aggregated.mut() function}
}
\description{
Compute statistics related to somatic selection
}
\examples{
: compute.selection.stats(output.folder='./examples/', output.prefix='TCGA.ACC');
}
