% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{compute.spatial.stats}
\alias{compute.spatial.stats}
\title{Compute statistics related to mutational spatial distributions}
\usage{
compute.spatial.stats(output.folder, output.prefix, bw = 5, m = 3)
}
\arguments{
\item{output.folder:}{A folder containing a "mut.summary.txt" file produced by the parse.aggregated.mut() function.}

\item{output.prefix:}{A prefix string used label the various files. Spatial statistics is written to a file "prefix.spatial.txt".}
}
\description{
Compute statistics related to mutational spatial distributions
}
\examples{
: compute.spatial.stat(folder='./examples/', prefix='TCGA.ACC');
}
