% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{correct.small.sample.bias}
\alias{correct.small.sample.bias}
\title{Correct biases for genes with too few somatic variants.}
\usage{
correct.small.sample.bias(opts.scs)
}
\arguments{
\item{opts.scs:}{A data frame containing all features.}
}
\value{
A data frame containing all features after correction
}
\description{
Correct biases for genes with too few somatic variants.
}
