% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{find_peaks}
\alias{find_peaks}
\title{Detect mutational hotspots}
\usage{
find_peaks(x, m = 3, bw = bw)
}
\arguments{
\item{x:}{a vector of mutated positions}
}
\value{
A data frame with peak coordinates
}
\description{
Detect mutational hotspots
}
