% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{get.exp}
\alias{get.exp}
\title{Get expected counts of variants for a vector of gene symbols}
\usage{
get.exp(symbols.focus)
}
\arguments{
\item{symbols.focus:}{A vector of gene symbols}
}
\value{
A data frame with expected counts of variants
}
\description{
Get expected counts of variants for a vector of gene symbols
}
