% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{get.opt.genes}
\alias{get.opt.genes}
\title{Estimate selection coefficients}
\usage{
get.opt.genes(data.all, genes.opt.file.name = "", indel = FALSE, map)
}
\arguments{
\item{data.all:}{A data frame with observed counts of variants}

\item{genes.opt.file.name:}{A file to write the computed selection coefficients.}

\item{map:}{A data frame containing symbol_2_cds_id mapping}
}
\value{
A data frame with selection coefficients
}
\description{
Estimate selection coefficients
}
