% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{gust}
\alias{gust}
\title{Perform GUST classification of oncogens, tumor suppressor genes and passenger genes from exome sequencing data.}
\usage{
gust(input.file.name, output.folder, output.prefix, steps = 1:8)
}
\arguments{
\item{input.file.name:}{A VCF-formatted file to read SnpEff annotated somatic variants. This input file shall contain these fields: Tumor_Sample_Barcode, Chromosome, Start_Position, dbSNP_RS, Reference_Allele, Tumor_Seq_Allele2, FILTER, One_Consequence, Hugo_Symbol, Gene, Feature, ENSP, HGVSc, HGVSp_Short, Amino_acids, Codons, ENSP, RefSeq, Entrez_Gene_Id.}

\item{output.folder:}{the folder to write the temporary and final prediction files.}

\item{output.prefix:}{prefix used to name the temporary and final prediction files. Temporary files include prefix.outlier.txt, prefix.mut.cnt.txt, prefix.mut.filtered.txt, prefix.mut.summary.txt, prefix.error.txt, prefix.log.txt, prefix.symbol_2_cds_id.txt, prefix.genes.opt.txt, prefix.spatial.txt, prefix.fitch.txt, prefix.combined.txt, prefix.reformatted.txt, prefix.gust.txt}

\item{steps:}{A vector of integers indicating which functions to execute. 1-find.outliers(), 2-parse.aggregated.mut(), 3-compute.selection.stats, 4-compute.spatial.stats(), 5-compute.fitch.stats(), 6-append.info(), 7-reformat(), 8-make.prediction(). Default to 1:8.}
}
\description{
Perform GUST classification of oncogens, tumor suppressor genes and passenger genes from exome sequencing data.
}
\examples{
gust(input.file.name='./examples/TCGA.ACC.mutect.somatic.maf.gz', output.folder='./examples/', output.prefix='TCGA.ACC');
}
