% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{initialize.gust}
\alias{initialize.gust}
\title{Prepare data for GUST classification
return a list of data frames, containing fitch rates of protein positions.}
\usage{
initialize.gust()
}
\description{
Prepare data for GUST classification
return a list of data frames, containing fitch rates of protein positions.
}
\examples{
initialize.gust();
}
