% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{make.prediction}
\alias{make.prediction}
\title{Predict oncogenes, tumor suppressor genes and passenger genes.}
\usage{
make.prediction(output.folder, output.prefix)
}
\arguments{
\item{output.folder:}{A folder containing a "reformatted.txt" file.}

\item{output.prefix:}{A prefix string used to label the files. Predictions are written to a file "prefix.predictions.txt".}
}
\description{
Predict oncogenes, tumor suppressor genes and passenger genes.
}
\examples{
: make.prediction(output.folder='./examples/', output.prefix='TCGA.ACC');
}
