% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{organize.obs}
\alias{organize.obs}
\title{Organize a collection of variants in a gene into counts}
\usage{
organize.obs(x)
}
\arguments{
\item{x:}{a data frame with four columns: name, type, mut_type, cnt.}
}
\value{
: a data frame with counts of variants grouped by types
}
\description{
Organize a collection of variants in a gene into counts
}
