% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.plot.R
\name{plot.multiple}
\alias{plot.multiple}
\title{Plot distribution of somatic mutations of multiple genes.}
\usage{
\method{plot}{multiple}(multiple, par = F, bw = 5, side = 1)
}
\arguments{
\item{multiple:}{A data frame. Each row contains a gene symbol, a folder, a prefix and an optional class label.}

\item{par:}{A boolean indicating if using current grid arrangement. Default is TRUE.}

\item{bw:}{bandwidth for density estimate. Default is 5.}

\item{main:}{title of the plot.}

\item{side:}{side to plot the protein position.}
}
\value{
A data frame from the check.gene.mut() function
}
\description{
Plot distribution of somatic mutations of multiple genes.
}
\examples{
df=data.frame(rbind(c('TP53', './examples', 'TCGA.ACC', 'TSG'), c('CTNNB1', './examples/', 'TCGA.ACC', 'OG'), c('TTN', './examples/', 'TCGA.ACC', 'PG'))); m=plot.multiple(df)
}
