% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.plot.R
\name{plot.mut}
\alias{plot.mut}
\title{Plot distribution of somatic mutations.}
\usage{
\method{plot}{mut}(mut.in.tumor, par = T, bw = 5, main = "",
  side = 1)
}
\arguments{
\item{mut.in.tumor:}{A list returned from the check.gene.mut() function.}

\item{par:}{A boolean indicating if using current grid arrangement. Default is TRUE.}

\item{bw:}{bandwidth for density estimate. Default is 5.}

\item{main:}{title of the plot.}

\item{side:}{side to plot the protein position.}
}
\description{
Plot distribution of somatic mutations.
}
