% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.plot.R
\name{plot.this.one}
\alias{plot.this.one}
\title{Plot distribution of somatic mutations of a gene.}
\usage{
\method{plot}{this.one}(symbol, folder, prefix, par = F, bw = 5,
  main = "", side = 1)
}
\arguments{
\item{symbol:}{A gene symbol.}

\item{folder:}{A folder containing a "prefix.mut.summary.txt" file and a "prefix.symbol_2_cds_id.txt" file.}

\item{prefix:}{A prefix string used to label the files. Usually, it contains TCGA.tumor_abbreviation.}

\item{par:}{A boolean indicating if using current grid arrangement. Default is TRUE.}

\item{bw:}{bandwidth for density estimate. Default is 5.}

\item{main:}{title of the plot.}

\item{side:}{side to plot the protein position.}
}
\value{
A data frame from the check.gene.mut() function
}
\description{
Plot distribution of somatic mutations of a gene.
}
\examples{
m=plot.this.one('TP53', './examples/', 'TCGA.ACC')
}
