% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{prob.joint.indel}
\alias{prob.joint.indel}
\title{optimization function to estimate joint probabilities of different types of variants including indels.}
\usage{
prob.joint.indel(par, obs_exp)
}
\description{
optimization function to estimate joint probabilities of different types of variants including indels.
}
