% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{prob.joint}
\alias{prob.joint}
\title{optimization function to estimate joint probabilities of different types of variants. Indels are treated independently from substitutions.}
\usage{
prob.joint(par, obs_exp)
}
\description{
optimization function to estimate joint probabilities of different types of variants. Indels are treated independently from substitutions.
}
