% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{reformat}
\alias{reformat}
\title{Reformat a data frame to be suitable for modeling.}
\usage{
reformat(output.folder, output.prefix)
}
\arguments{
\item{output.folder:}{A folder containing a "combined.txt" file.}

\item{output.prefix:}{A prefix string used to label the files. Reformatted data frame is written to a file "prefix.reformatted.txt".}
}
\value{
A data frame containing all features after reformat
}
\description{
Reformat a data frame to be suitable for modeling.
}
\examples{
: reformat(output.folder='./examples/', output.prefix='TCGA.ACC');
}
